+++ click_and_drop_api/simple/__init__.py
@@ -21,18 +21,12 @@ from .types import (
 )
 from .api import ClickAndDrop
 from .mock import MockClickAndDrop
-from .package_sizes import (
-    PackageSize,
-    packages_sizes,
-    get_package_size,
-    choose_package_size_by_weight,
-    get_package_sizes,
+from .shipping.db import (
+    ShippingDB,
+    PackageShippingOption,
+    db,
 )
-from .shipping_options import (
-    ShippingOption,
-    list_service_codes,
-    check_service_codes,
-)
+from .shipping import check_service_codes
 from .errors import InvalidWeight, InvalidDimensions
 
 __all__ = [
@@ -43,14 +37,10 @@ __all__ = [
     "InvalidDimensions",
     "check_service_codes",
     "RecipientDetails",
-    "list_service_codes",
     "Address",
-    "get_package_sizes",
-    "PackageSize",
-    "packages_sizes",
-    "ShippingOption",
-    "choose_package_size_by_weight",
-    "get_package_size",
+    "ShippingDB",
+    "PackageShippingOption",
+    "db",
     "CreateOrders",
     "UpdateOrdersStatus",
     "UpdateOrderStatus",